update assertions
- create new `MakesUrlAssertions` trait and pull out all of the URL assertions into this file, and remove them from `MakesAssertions`
- use the new trait in the `Browser`
- add new `assertSchemeIs()` method to `MakesUrlAssertions` to check the scheme of the URL
- separate assertions tests into new `MakesUrlAssertionsTest` and pull out all the URL assertions tests from `MakesAssertionsTest`
- add test for new `assertSchemeIs()` method
